delete notes

+++ notes-app/src/main/java/net/project/springboot/models/User.java
@@ -13,7 +13,7 @@ import javax.persistence.OneToMany;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
 
     private String username;
     private String password;
@@ -29,11 +29,11 @@ public class User {
     // this.notes = notes;
     // }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
